This PR is a part of a series of PRs to enable chunked prefill in aphrodite, mainly to save vram for large context sizes, and possibly speedup throughput. vLLM is currently employing a similar technique, so we will expand upon that. AFAIK, exllamav2 also has an efficient chunked decoding implementation.

Part 1 (here) addresses the shape of the query, changing it from 2D to 1D.
Part 2 will update the seqgroup metadata API.
Part 3 will need us to refactor the scheduler, by moving the logic for prefill scheduling and decoding scheduling to a dedicated function. We want a different scheduling policy for chunked prefill. Currently, we do prefill -> decoding, but when chunked, it should be decoding -> prefill.
Part 4 will add a separate scheduler for the chunked prefill
Part 5 will see it enabled in the attention code