@sorasoras 2024-06-01
> Outputs are virtually unchanged (confirmed with seeded gens) and there's a bit of latency reduction.
> 
> 2x A40s, Mistral 7B Instruct v0.2:
> 
> PR:
> 
> ```
> Successful requests: 200
> Benchmark duration: 69.589456 s
> Total input tokens: 45025
> Total generated tokens: 44697
> Request throughput: 2.87 requests/s
> Input token throughput: 647.01 tokens/s
> Output token throughput: 642.30 tokens/s
> Mean TTFT: 213.36 ms
> Median TTFT: 165.27 ms
> P99 TTFT: 771.10 ms
> Mean TPOT: 165.13 ms
> Median TPOT: 129.63 ms
> P99 TPOT: 514.82 ms
> ```
> 
> Main:
> 
> ```
> Successful requests: 200
> Benchmark duration: 69.714347 s
> Total input tokens: 45025
> Total generated tokens: 44579
> Request throughput: 2.87 requests/s
> Input token throughput: 645.85 tokens/s
> Output token throughput: 639.45 tokens/s
> Mean TTFT: 352.50 ms
> Median TTFT: 305.21 ms
> P99 TTFT: 872.61 ms
> Mean TPOT: 207.33 ms
> Median TPOT: 138.11 ms
> P99 TPOT: 787.68 ms
> ```

How much saving on vram in long context compare to not using it on a gqa model  on 16K context for example ？